Efficiency, Comments and OOB-Bug updates
I've noticed that the RedBalls have a constant way of moving (or at least, kind of), so I've decided to save that thing in a constant for each RedBall for efficiency.  
I've then put tons of comments everywhere because everyone was asking what the code meant, now it should be clear enough.  
The "out of bounds" bug is resolved, there's still the "I'm unable to move because I'm between two BlueBalls" bug but actually it's what you get for standing still in the middle of these eheh